capstoneproject/backend/gateway/requirements.txt
flask==3.0.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
requests==2.32.3
pyjwt==2.8.0
